+++ client/a8c-for-agencies/sections/referrals/referral-details/index.tsx
@@ -1,16 +1,14 @@
 import { useDesktopBreakpoint } from '@automattic/viewport-react';
 import { useTranslate } from 'i18n-calypso';
 import { useMemo, useState } from 'react';
-import ItemView, {
-	createFeaturePreview,
-} from 'calypso/layout/multi-sites-dashboard/item-preview-pane';
+import ItemView, { createFeaturePreview } from 'calypso/layout/multi-sites-dashboard/item-view';
 import SubscriptionStatus from '../referrals-list/subscription-status';
 import ReferralCommissions from './commissions';
 import ArchivedStatus from './components/archived-status';
 import ReferralPurchasesMobile from './mobile/purchases-mobile';
 import ReferralPurchases from './purchases';
 import type { Referral, ReferralInvoice } from '../types';
-import type { ItemData } from 'calypso/layout/multi-sites-dashboard/item-preview-pane/types';
+import type { ItemData } from 'calypso/layout/multi-sites-dashboard/item-view/types';
 
 import './style.scss';
