+++ client/a8c-for-agencies/sections/sites/features/jetpack/jetpack-preview-pane.tsx
@@ -11,10 +11,8 @@ import {
 } from 'calypso/a8c-for-agencies/sections/sites/features/features';
 import SitesDashboardContext from 'calypso/a8c-for-agencies/sections/sites/sites-dashboard-context';
 import { useJetpackAgencyDashboardRecordTrackEvent } from 'calypso/jetpack-cloud/sections/agency-dashboard/hooks';
-import ItemView, {
-	createFeaturePreview,
-} from 'calypso/layout/multi-sites-dashboard/item-preview-pane';
-import { ItemData } from 'calypso/layout/multi-sites-dashboard/item-preview-pane/types';
+import ItemView, { createFeaturePreview } from 'calypso/layout/multi-sites-dashboard/item-view';
+import { ItemData } from 'calypso/layout/multi-sites-dashboard/item-view/types';
 import { PreviewPaneProps } from '../../site-preview-pane/types';
 import { JetpackActivityPreview } from './activity';
 import { JetpackBackupPreview } from './backup';
