+++ client/layout/multi-sites-dashboard/item-view/item-view-content/style.scss
